Create the LUSTRE FSAL over VFS FSAL
This initial Lustre FSAL code is as nearest as possible from the
VFS FSAL code.

It only adds an HSM check to trigger an hsm restore and return
ERR_FSAL_DELAY when a file is released. An option of export is
added to set/unset this behaviour : async_hsm_restore. For
example, 9P clients don't support retry.

If we don't have an up to date lustre api, we still compile a
"fake" LUSTRE FSAL to continue to check the compilation path.

Change-Id: Ibe495fd0467ab6b86c1eb369099a78fc048d55a0
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>