support: add a rados_grace support library
Add a new support library for coordinating a grace period between
multiple nodes. The basic idea is to use a shared RADOS object to
mediate requests for a grace period.

Typically, each node will perform a read operation on the database
and (then possibly) a write operation. In order to ensure atomicity,
we assert that the object version has not changed between the read
and write ops. If it has, then we start over again from scratch.

Change-Id: Idfff61f8d0092fd9095dc14390dfb434f0ad6696
Signed-off-by: Jeff Layton <jlayton@redhat.com>