Update patch set 2
Patch Set 2:

> (1 comment)
I think it has nothing to do with ACL expiration. We should carry out similar steps to that block where we do not get any ACLs << } else if (attrs->acl != NULL) { >>
To copy the ACLs from attr-> to entry->attr.acl, we need to have that flag set irrespective whether we had ACLs or not.

Patch-set: 2