SAL: don't revoke delegations on export release
This reverts most of commit d002f04fa8f0 (Revoke delegation states when
an export is released).

The problem there is that we call into the code to release exports on
a clean shutdown, and we most certainly do not want to revoke a
delegation just for that.

I'm not sure of the rationale for that commit, as releasing an export
doesn't mean that any client of it is necessarily delinquent in
returning a delegation. It may never have ever been recalled.

Just have the code release the stateid in that case like it did before.

We call into the code to release exports on server shutdown, so

Change-Id: Ic2a4a4e3d9609e87ce821e02764047de913415cd
Cc: Malahal Naineni <malahal@us.ibm.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>