RADOS_KV: make rados_kv_traverse callback more generic
We have a struct to package all the arguments, but then we pass all of
the arguments individually. Have the rados_kv_traverse callback take a
void pointer and then just have the callback cast it to the right sort
of struct to get the arguments.

While we're in there, also get rid of the useless pop_args_t typedef.

This is mainly just a cleanup to untangle the API a bit, but it also
allows us to potentially pass new and different sorts of callbacks to
rados_kv_traverse.

Change-Id: Ib9141a79313982577a41b831630d4166740c1a86
Signed-off-by: Jeff Layton <jlayton@redhat.com>