FSAL_CEPH: fix libcephfs result handling in read2/write2 ops
The recent API changes to add vectored I/O to FSAL_CEPH don't handle
the result correctly:

First, it breaks out of the loop before fixing up the lengths if the
offset is 0. That means that reading from the beginning of the file
returns a bunch of NULLs and writes to the beginning of the file fail.

Also, it seems to place some significance on the offset being -1. I'm
unclear on why we'd treat that as a special case. Either it was already
-1ULL on entry to the function, or we will have just happened to have
hit that magic offset after iterating over a particular iovec. Let's
just remove that check and use a zero-length but successful read or
write as an indicator to stop looping.

Finally, treat a successful but zero-length read as an EOF condition,
instead of requiring that the client issue another READ to discover
that.

Change-Id: I628f0a694f1a0b26e9d76203a59e4493c0a8e25e
Reported-by: Patrick Donnelly <pdonnell@redhat.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>