SAL: add recovery operation to maybe start a grace period
Add a new maybe_start_grace operation that will check to see if a
sibling has requested a grace period and start one if so. Call this
function from the reaper thread.

Note that we only enforce the grace period and do not reload the
recovery db.  All of the current node's state is still intact, so we
don't need to initiate any reclaim; only enforce the grace period until
the other node's reclaim is complete.

Change-Id: I85f54907efb220510af0b87baa1798d3f1d7a041
Signed-off-by: Jeff Layton <jlayton@redhat.com>