Fix strncmp() size argument.
For a comparision with a string literal, say "PSEUDO", the max bytes to
compare should be sizeof("PSEUDO") or strlen("PSEUDO")+1 unless, of
course, we are only trying to make a prefix comparision.

If one of the strings to strncmp is a string literal, then there is no
advantage using strncmp at all, and we could just use strcmp() as well.

Change-Id: Ib1c57aa1db76c0ae98a8f37a41daf55ed2d7c7b2
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>