Add new CMake option code
CMake options were either on or off, and no way to tell if they were
using the default value or if they were specified on the command line.
This made it extremely difficult to error out if something was specified
on the command line but the dependencies were not installed.

Add a new pair of macros for option processing:
- goption() is a replacement for option(), and it records the default
  value specially
- gopt_check() should be called before using the option, and it
  determins whether or not the default was used, and sets <foo>_REQUIRED
  as a result.  This can be used to error out if the option was given on
  the command line.

Most options are converted to this new API.

Change-Id: I83609375ddc27648bc6eb2bb9fe73848f571a309
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>