Update patch set 1
Patch Set 1:

Hmm, I'm still not sure all the possible combinations are handled...

Some bits and bobs:

If we get a replayed exclusive create, MDCACHE may call open2 with an obj_handle instead of a name with createmode set. The FSAL then is responsible for checking the verifier.

If we get an UNCHECKED create, MDCACHE may call open2 with an obj_handle instead of a name with createmode set.

In either of these cases, I think the intent is that if indeed the file does not exist, ENOENT is returned, and then MDCACHE will actually call open2 again with the name.

EXCLUSIVE create should always succeed if the file doesn't exist or the file exists but has the proper verifier though there COULD be a race where the file is actually deleted too late for exclusive create to succeed due to the timing of an unlink, but we shouldn't fail just because the MDCACHE wasn't correctly in sync.

UNCHECKED create should always succeed baring there being no room to create more files or some hard error in the filesystem.

GUARDED create may fail and the file end up actually deleted. If a GUARDED create races with an unlink we should at least try to cover for the case where the MDCACHE is out of sync with the filesystem.

The MDCACHE code might not be quite right... (it probably should retry the creates by name if ESTALE comes back for the obj_handle...).

Patch-set: 1