main: add way to stall server until grace is being enforced
To prevent conflicting state being handed out before you do reclaim, we
need to wait until everyone is in enforcing mode before we ask the MDS
to tear down any old sessions that may be present.

The simplest way to do this is to delay startup of the server until
everyone else is enforcing. Note that a singleton server should never
end up waiting here.

Change-Id: I0af053a3e2c91ee72ee6a0a0a9474d411c553e97
Signed-off-by: Jeff Layton <jlayton@redhat.com>