FSAL_CEPH: kill off old session before the mount
Add a test for the new cephfs reclaim infrastructure. We don't actually
try to reclaim at this point, only ask the MDS to destroy the old
session. For now, we just use the hostname as the uuid here.

The libcephfs interface is still under debate, so this may look quite
different in its final form.

Change-Id: I70e93cbff2e2ae80989aaa9ee9c96d3031f5e919
Signed-off-by: Jeff Layton <jlayton@redhat.com>