CMake - make goption work with cache
goption was saving the ON/OFF status (as opposed to the quad-state) in
the cache, breaking re-runs of CMake.

Fix goption to only store ON/OFF state in memory, and cache the correct
quadstate.

Also, remove the extra cache set calls that Ganesha had.  They're not
necessary anymore, goption takes care of it.

Change-Id: I4aa80a4a9ade3500f01f19338a521362d07d8026
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>