Update patch set 1
Patch Set 1: Code-Review-1

No, this is intentional.  It indicates whether or not we trust the cache for a lookup, or whether it must be invalidated first.  readdir() will populate regardless of this flag (it operates on the MDCACHE_DIR_POPULATED flag).

Patch-set: 1
Reviewer: Daniel Gryniewicz <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1