Update patch set 1
Patch Set 1: Code-Review-1

For non-FSAL_PSEUDO pseudofs root exports, I see no problem with this.

There is an issue for FSAL_PSEUDO in that it's handles are too large for NFSv3. This we should verify in the FSAL_PSEUDO export processing that it is not exported NFSv3 and if so, return an error.

Optionally we could do more work and if an FSAL_PSEUDO export is exported to NFSv3 then FSAL_PSEUDO should reduce it's handle size to 64 bytes. That will mean more situations where the file handle is mashed up (FSAL_PSEUDO uses the Pseudo path as the handle if it fits in 128 bytes, otherwise it creates a hash value, this is so it can generate the same file handles each time Ganesha starts with different exports).

Patch-set: 1
Reviewer: Frank Filz <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1