packaging: add cmake and BuildRequires for MSPAC support
While trying to build FSAL_CEPH and/or FSAL_RGW for nfs-ganesha-2.5.5 in
Fedora 27, discovered that WBCLIENT_LIBRARIES is not found by cmake, i.e.
missing MSPAC support and libwbclient-devel. (Unclear why this is only
found when building the Ceph FSALs; without those enabled it builds fine.)

jlayton's recent fix to FindWBclient.cmake in change
  Ie1ba82261e1728f300037127d3efd3f22a70ebfe
allows cmake to silently disable MSPAC when wbclient(-devel) is not
installed. But that patch hasn't been backported to 2.5-stable (yet).

Independent of that, conditionalize adding the BuildRequires and
configuring when MSPAC support is requested.

Change-Id: I8fb5abb278bbddd0ee8f93924737891c485a527f
Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>