Update patch set 3
Patch Set 3:

> I think you want to wrap that inside a if(_MSPAC_SUPPORT)
> conditional like the RPCBIND one above it?

Sure, but I was trying to have it always default to off in the rpmbuild.
Wrapped in if/else/endif it looks to me like it tests if wbclient-devel is installed, if so, enable it, and then make generate the .spec file such that wbclient-devel is installed or not with the BuildRequires.

But maybe I'm imagining that. I'll add the if/else/endif

Patch-set: 3
Reviewer: kaleb@redhat.com <1004389@d5d70762-12d0-45a1-890d-524b12d3f735>