Update patch set 1
Patch Set 1:

There is some question as to whether we should make the spec and the pynfs test less restrictive, and allow an ENOTEMPTY return in this situation as well. I'd be fine with that, but given that even v3 has a similar requirement and that we're not likely to be able to rev that spec, we should probably ensure that we return EEXIST in this situation.

One alternative here would be to catch this error at a higher level (maybe in fsal_rename?), which would fix this for all FSALs rather than just FSAL_CEPH.

Patch-set: 1
Reviewer: Jeff Layton <1008909@d5d70762-12d0-45a1-890d-524b12d3f735>