Use euid/egid instead of fsuid/fsgid
Doing this, we can then call getuid and getgid for saving the
credentials. This allows the interfaces for
fsal_save_ganesha_credentials to be a bit more generic.

Since we want to set thread specific credentials, we call the system
call directly rather than using setresuid/setresgid which is are glibc
functions that cause the system calls to be made on each thread in the
process.

Change-Id: Iaedfb0b9875ca0610171652bd160d748ed429a73
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>