SAL: make recovery_init operation an int return
If the recovery_init operation fails, then we're likely to end up
crashing later anyway as certain pointers we rely on will end up being
NULL, etc.

Failure to initialize the recovery backend should really be a fatal
startup error that prevents the server from starting. Make recovery_init
an int return operation and have the server die if it fails with an
error.

Change-Id: I3fb8ce57e7715801cc069970c782a26d4a61d894
Signed-off-by: Jeff Layton <jlayton@redhat.com>