Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Why can't the first machine started create the DB?

That's what happens now, but we really want adding a new cluster node to be a deliberate act. We want an administrator (or orchestration infrastructure) to say: add host "foo" to the cluster.

That's done by adding an omap key for host "foo". Since you have to add an omap key before the host can do anything, the db must already exist. So, creating the DB on startup is not something that we really want each server to do.

Patch-set: 2
Reviewer: Jeff Layton <1008909@d5d70762-12d0-45a1-890d-524b12d3f735>