Update patch set 2
Patch Set 2:

Daniel, could you try this version.

I think there still are some holes, places where we free dirents that may not be properly accounted for. I think the most likely hole is in
the re-use of an existing dirent.

I considered adjusting dirents_used where allocated and freed (I added dirents_allocated as a check), but the convoluted ways that an allocated dirent may or may not be put into a chunk made it hard to find the right way to reserve dirents_used on allocation of a chunk yet, result in the proper number of dirents_used at the end.

Patch-set: 2
Reviewer: Frank Filz <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>