Update patch set 5
Patch Set 5:

(1 comment)

> None of this is VFS specific.  Should we just put the option in
 > fsal_module and change fsal_set_credentials()?  It'd be much
 > cleaner.

Or even, I wonder if it really should be a global option, and the EPERM for wrong user credentials could just be handled up in the protocol layer when we handle credentials? Or it could be an export option (maybe tagged onto Squash as Squash=Single_User; or something like that).

Patch-set: 5
CC: Frank Filz <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>