specfile: move ganesha-rados-grace into a separate package
This is not really associated with FSAL_CEPH, so packaging it there is
probably not right.

Rename nfs-ganesha-rados package to nfs-ganesha-rados-grace. The old
package is not built currently because there are no files in it, and
the new name better describes what will be inside it.

Also add in the ganesha-rados-cluster-design manpage which wasn't being
packaged before.

Change-Id: I7f607b0fea7a720aa0c8a6be31bc18e1caeb68a7
Signed-off-by: Jeff Layton <jlayton@redhat.com>