Update patch set 4
Patch Set 4: Code-Review+1

It looks like the original plan was to have the backends be loadable libraries, like FSALs are, and so a separate package was made.  Work was never done to make them loadable modules, so this is fine for now.

Patch-set: 4
Reviewer: Daniel Gryniewicz <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1