exports: release export state after calling ->unexport op
Nothing defines the "unexport" operation today so this change should not
be noticeable at all. In the future though, we want to allow FSAL_CEPH
to prepare an export for clean shutdown such that we don't tear down
persistent state when tearing down objects.

We can use the unexport operation for that, but that means that we need
to do it before calling state_release_export.

Change-Id: I8edecc24d7e70dcb64eac9f5284d141a2b040bc9
Signed-off-by: Jeff Layton <jlayton@redhat.com>