Update patch set 1
Patch Set 1: Code-Review-1

So, I think this is not the way to fix this.  The reason is that normal locks won't work, since they are released on ever close (which we do all the time).  Coding it this way will result in the code compiling and appearing to run, but on locks being lost all the time, if OFD is not available.

A better solution, I think, is to conditionalize the setting of lock_op2 for VFS in FSAL_VFS/handle.c:1693 with the same ifdef.  It will then inherit the default version of the op, which just returns ENOTSUP.

Patch-set: 1
Reviewer: Daniel Gryniewicz <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1