Update patch set 1
Patch Set 1:

This will break XDEV cases.  That is, you can never mount a different filesystem (either FSAL_VFS or other FSAL) anywhere in a VFS tree.  This is quite limiting, I think, and probably should be addressed.

How about this as a solution:  Add an extra field on the beginning of v_fid that is internal to Ganesha, and use that in place of fid_reserved in all the changes.  Then, when you pass v_fid into any kernel calls, pass the offset skipping that field (this can be done with a macro, to it's always done right).  The kernel will never see that field, and we can store whatever we want in it.

Patch-set: 1
CC: Daniel Gryniewicz <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>