Update patch set 1
Patch Set 1:

> This will break XDEV cases.  That is, you can never mount a
 > different filesystem (either FSAL_VFS or other FSAL) anywhere in a
 > VFS tree.  This is quite limiting, I think, and probably should be
 > addressed.
 > 
 > How about this as a solution:  Add an extra field on the beginning
 > of v_fid that is internal to Ganesha, and use that in place of
 > fid_reserved in all the changes.  Then, when you pass v_fid into
 > any kernel calls, pass the offset skipping that field (this can be
 > done with a macro, to it's always done right).  The kernel will
 > never see that field, and we can store whatever we want in it.

This will only cause breakage if a FreeBSD setup actually has a VFS export with another FSAL handling a sub-directory of that VFS export, so I'm not too worried about it.

That said, considering a solution would be good.

Are there any values of fid_reserved that would never be seen? Note that these dummy handles are ONLY used to instantiate an MDCACHE entry in the VFS export to serve as a junction for the sub-mounted export.

In Linux land, it comes into play for a VFS->XFS, XFS->VFS, VFS->GPFS, or XFS->GPFS boundary (I don't think FSAL_GPFS produces dummy handles, so GPFS->VFS or GPFS->XFS boundary is not supported).

Patch-set: 1
CC: Frank Filz <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>