FSAL_VFS: fix read_dirents on FreeBSD
The dirent structure on FreeBSD does not implement the d_off field.
There's some computation done in to_vfs_dirent for FreeBSD but it's not
valid for upstream filesystems. We could try to compute offsets manually
but this is painful, especially with ZFS.
A workaround to retrieve offsets is to rewind and re-read the directory
entry by entry and fetch each offset.
This is very inefficient as it will do a syscall for each entry.
We also have to reduce the buffer size for batched reads
in order to narrow races due to the rewind process.

FreeBSD 12 introduced the d_off field but is currently unused.
We can use it once it is fully available

Change-Id: I75416c6981b921b63f9d493854ece41260fca157
Signed-off-by: Fatih Acar <fatih.acar@gandi.net>