FSAL: Convert fsal_obj_ops in fsal_obj_handle to pointer
This trades an extra dereference for a dramatic reduction in the size of
the fsal_obj_handle (which with stacked mdcache is two fsal_obj_handle
structures per cached inode).

The extra dereference will dramatically reduce memory consumtion
(424 bytes per fsal_obj_handle * 2 * 100,000 is 85 MB).

We can also expect some benefit to having a single function vector
per FSAL resulting in the function vector often liing in L2 cache.
A further optimization (coming in a subsequent patch) will be to
move the most used methods to the front of the structure in the hopes
that they can live in a single cache line and this may even live in
the L1 cache on the CPU.

Change-Id: Ie20e50900488d72e7871f2be295cb8241fbc0f3f
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>