Update patch set 1
Patch Set 1:

> So, looking at the code, in do_lock_op(), holder is only set (and
 > only has it's ref incremented) if status is STATE_LOCK_CONFLICT,
 > not if status is STATE_LOCK_BLOCKED, so this holder will always be
 > NULL, so this code will never trigger.

There is also a call to copy_conflict in state_lock() if it finds a conflicting lock.

I'm pretty sure we can actually solve this just by not making that call to copy_conflict if it's a blocking lock (since we will be returning STATE_LOCK_CONFLICT and this holder and conflict will not be used.

Patch-set: 1
Reviewer: Frank Filz <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>