Update patch set 1
Patch Set 1:

Hmm, 1 and 2 may not be reserved in the kernel. EXT4 uses the same hash it uses for named entries, and thus they show up in the middle of a directory (do an ls -aU1 to see for yourself...).

Patch-set: 1
Reviewer: Frank Filz <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>